Make versioning consistent for avro and json
