+++ core-common/src/test/java/org/zalando/nakadi/domain/VersionTest.java
@@ -8,7 +8,7 @@ import static org.junit.Assert.assertThat;
 public class VersionTest {
     @Test
     public void bumpShouldZeroSmallerVersionNumber() throws Exception {
-        assertThat(new JsonVersion("1.1.0").bump(JsonVersion.Level.MAJOR), equalTo(new JsonVersion("2.0.0")));
-        assertThat(new JsonVersion("1.1.1").bump(JsonVersion.Level.MINOR), equalTo(new JsonVersion("1.2.0")));
+        assertThat(new Version("1.1.0").bump(Version.Level.MAJOR), equalTo(new Version("2.0.0")));
+        assertThat(new Version("1.1.1").bump(Version.Level.MINOR), equalTo(new Version("1.2.0")));
     }
 }
