+++ core-metastore/src/main/java/org/zalando/nakadi/repository/db/EventTypeRepository.java
@@ -12,11 +12,9 @@ import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.RowMapper;
 import org.springframework.stereotype.Component;
 import org.zalando.nakadi.annotations.DB;
-import org.zalando.nakadi.domain.AvroVersion;
+import org.zalando.nakadi.domain.Version;
 import org.zalando.nakadi.domain.EventType;
 import org.zalando.nakadi.domain.EventTypeBase;
-import org.zalando.nakadi.domain.EventTypeSchemaBase;
-import org.zalando.nakadi.domain.JsonVersion;
 import org.zalando.nakadi.exceptions.runtime.DuplicatedEventTypeNameException;
 import org.zalando.nakadi.exceptions.runtime.InternalNakadiException;
 import org.zalando.nakadi.exceptions.runtime.NoSuchEventTypeException;
@@ -42,7 +40,7 @@ public class EventTypeRepository extends AbstractDbRepository {
             DuplicatedEventTypeNameException {
         try {
             final DateTime now = new DateTime(DateTimeZone.UTC);
-            final var version = getStartVersion(eventTypeBase.getSchema().getType());
+            final var version = getStartVersion();
             final EventType eventType = new EventType(eventTypeBase, version, now, now);
             jdbcTemplate.update(
                     "INSERT INTO zn_data.event_type (et_name, et_event_type_object) VALUES (?, ?::jsonb)",
@@ -143,9 +141,8 @@ public class EventTypeRepository extends AbstractDbRepository {
         }
     }
 
-    private String getStartVersion(final EventTypeSchemaBase.Type type){
-        return type == EventTypeSchemaBase.Type.JSON_SCHEMA?
-                new JsonVersion().toString(): new AvroVersion().toString();
+    private String getStartVersion(){
+        return new Version().toString();
     }
 
     public enum RowLockMode {
